requirements.txt
Fabric==1.7.0
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
ecdsa==0.10
httplib2==0.8
itsdangerous==0.23
paramiko==1.12.0
pycrypto==2.6.1
python-instagram==0.8.0
simplejson==3.3.1
wsgiref==0.1.2